requirements.txt
gradio==4.25.0
diffusers>=0.30.0
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
safetensors>=0.4.0
pillow>=10.0.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
